% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr.hoeffding}
\alias{corr.hoeffding}
\title{Hoeffding's D and Spearman matrix in order to produce a scatterplot
Inspired in SAS Certification Prep Guide: Statistical Business Analysis Using SAS9}
\usage{
corr.hoeffding(x, y)
}
\arguments{
\item{x}{Matrix x}

\item{y}{Response y}
}
\value{
The matrix with the following columns:
  \item{spearman_corr_coeff}{Correlation coefficient for Spearmans}
  \item{hoeffding_d_corr_coeff}{Correlation coefficient for Hoeffding's D measure}
  \item{spearman_abs}{Absolute value of the correlation coefficient}
  \item{hoeffding_d_abs}{Absolute value of the Hoeffding's D correlation coefficient}
}
\description{
Hoeffding's D and Spearman matrix in order to produce a scatterplot
Inspired in SAS Certification Prep Guide: Statistical Business Analysis Using SAS9
}
\examples{

}
\seealso{
\link{https://www.ncbi.nlm.nih.gov/pubmed/23962479}
}
\keyword{hoeffding,}
\keyword{monotonic}
\keyword{spearman,}
