% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval.mcnemar}
\alias{eval.mcnemar}
\title{Compute McNemar's test}
\usage{
eval.mcnemar(
  response,
  model1,
  model2,
  names = c("Model 1", "Model 2"),
  headings = c("Correct", "Incorrect"),
  correct = TRUE
)
}
\description{
Compute McNemar's test
}
