% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{formula.build}
\alias{formula.build}
\title{Function to dynamically build formula}
\usage{
\method{formula}{build}(y, x = c(), regex = "",
  transformations = list())
}
\description{
Function to dynamically build formula
}
