% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trees.R
\name{simulate_null_tree}
\alias{simulate_null_tree}
\title{Simulate a constant-rate birth-death tree}
\usage{
simulate_null_tree(n, rho = 1, epsilon = 0)
}
\arguments{
\item{n}{Number of extant tips.}

\item{rho}{Sampling fraction.}

\item{epsilon}{Extinction fraction (mu = epsilon times lambda).}
}
\value{
An ultrametric \code{phylo} object.
}
\description{
Simulate a constant-rate birth-death tree
}
