% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{vectors.diff}
\alias{vectors.diff}
\title{Compute Vector Diff}
\usage{
vectors.diff(vector1, vector2)
}
\arguments{
\item{array1}{Vector to compare from}

\item{array2}{Vector to compare against}
}
\value{
Vector containing all the items from vector1 that are not part of vector2
}
\description{
This function computes the difference of arrays
}
